[FARM-279]-Circular simplified to only be a Atom, the CircularPainter file it's beed added in the CricularWidget file as a private class
