--- lib/ui/common/CircularPainter.dart
@@ -1,53 +1,0 @@
-
-
-import 'dart:math';
-
-import 'package:flutter/material.dart';
-
-class CircularPainter extends CustomPainter{
-  Color lineColor;
-  Color completeColor;
-  double completePercent;
-  double width;
-
-
-  CircularPainter({this.lineColor, this.completeColor, this.completePercent,
-      this.width});
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    Paint line = Paint()
-        ..color = lineColor
-        ..strokeCap = StrokeCap.round
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = width;
-    Paint complete = Paint()
-        ..color = completeColor
-        ..strokeCap = StrokeCap.round
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = width;
-
-    Offset center = Offset(size.width/2, size.height/2);
-    double radius = min(size.width/2, size.height/2);
-    canvas.drawCircle(
-        center,
-        radius,
-        line
-    );
-
-    double arcAngle = 2*pi* (completePercent/100);
-
-    canvas.drawArc(
-        Rect.fromCircle(center: center, radius: radius),
-        -pi/2,
-        arcAngle,
-        false,
-        complete
-    );
-  }
-
-  @override
-  bool shouldRepaint(CustomPainter oldDelegate) {
-    return true;
-  }
-}
